refactor: features/(authenticated)/post 구조 정리
