+++ src/app/(authenticated)/post/[id]/(detail)/page.tsx
@@ -12,7 +12,7 @@ import { getPostDetail, deletePost } from '@/features/(authenticated)/post/apis/
 import type { ReactionType } from '@/features/(authenticated)/post/types/Post.types'
 import { getPostDetailReaction } from '@/features/(authenticated)/post/apis/reaction.api'
 import { toggleReactionAction } from '@/features/(authenticated)/post/actions/toggleReaction.action'
-import { PostDetailHeader } from '@/features/(authenticated)/post/create/components/PostDetailHeader'
+import { PostDetailHeader } from '@/features/(authenticated)/post/components/PostDetailHeader'
 import { RelativeTime } from '@/components/RelativeTime'
 import { Avatar } from '@/components/ui/Avatar'
