+++ src/features/(authenticated)/post/utils/validatePost.ts
@@ -1,5 +1,5 @@
-import { TOPIC_TYPE } from '../../../../../types/Topic.types'
-import { PostFormValues } from '../../types/Post.types'
+import { TOPIC_TYPE } from '@/types/Topic.types'
+import { PostFormValues } from '@/features/(authenticated)/post/types/Post.types'
 
 export function validatePost(values: PostFormValues) {
   const fieldErrors: FormStateTypes<PostFormValues>['fieldErrors'] = {}
